pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "segmentation"
version = "0.1.0"
description = "Food segmentation application with FastAPI and Streamlit"
requires-python = ">=3.11"
dependencies = [
    "mkdocs>=1.6.1",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "prometheus-client>=0.22.1",
    "tqdm==4.67.1",
]

[tool.setuptools]
packages = ["src.app"]

[project.optional-dependencies]
backend = [
    "fastapi==0.115.12",
    "uvicorn==0.34.2",
    "python-multipart>=0.0.20",
    "torch==2.7.0",
    "torchvision==0.22.0",
    "scikit-learn==1.6.1",
    "matplotlib>=3.7.0",
    "omegaconf==2.3.0",
    "hydra-core==1.3.2",
    "loguru==0.7.3",
    "annotated-types==0.7.0",
    "celery==5.5.2",
    "dvc==3.59.1",
    "dvc-data==3.16.10",
    "dvc-task==0.40.2",
    "dvc-gs==3.0.2",
    "wandb==0.19.11",
    "gto==1.7.2",
    "dvc-studio-client>=0.21.0",
    "google-cloud-storage==3.1.0",
    "gcsfs==2025.5.1",
    "gdown==5.2.0",
    "prometheus-client>=0.22.1",
]

frontend = [
    "streamlit>=1.46.1",
    "requests>=2.31.0",
    "pillow>=10.0.0",
    "matplotlib",
]

dev = [
    # Your dev dependencies
]
docker = [
    "supervisor>=4.2.0",
]

[tool.coverage.run]
source = ["src.segmentation"]
omit = ["src/tests/**/*"]
